修改user_login_persistent为last_login为长整型

+++ src/main/java/com/hts/web/aliyun/dao/OsUserLoginCacheDao.java
@@ -4,9 +4,9 @@ import com.hts.web.common.dao.BaseCacheDao;
 
 public interface OsUserLoginCacheDao extends BaseCacheDao {
 
-	public void saveLogin(Integer id, Integer lastLogin);
+	public void saveLogin(Integer id, Long lastLogin);
 	
-	public void updateLogin(Integer id, Integer lastLogin);
+	public void updateLogin(Integer id, Long lastLogin);
 	
 	public void popOpts(int limit);
 }
